+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog of nenslint
 0.1 (unreleased)
 ----------------
 
+- Added check for proper ``.gitignore`` file.
+
 - Added automatic documentation generation.
 
 - Added txt-to-rst conversion checker.